Connected user verifications on user model

+++ backend/popay/src/main/kotlin/com/tdev/popay/controller/UserController.kt
@@ -56,44 +56,55 @@ class UserController(
         return ResponseEntity(ResponseMessage(false, "Invalid credentials"), HttpStatus.BAD_REQUEST)
     }
 
-    @GetMapping("/users")
-    fun getAllUsers(): List<User> = userService.findAll()
-
-    @GetMapping("/user/{id}")
-    fun getUserById(@PathVariable(value = "id") userId: Long): ResponseEntity<Any> {
-        val checkUser = userService.findById(userId)
-        if (checkUser != null) {
-            return ResponseEntity(checkUser, HttpStatus.OK)
+    @GetMapping("/user")
+    fun getUser(@RequestHeader("Authorization") token: String): ResponseEntity<Any> {
+        val tokenValue = token.substring(7)
+        val userId = tokenService.getUserIdFromToken(tokenValue)
+        if (userId != null) {
+            val checkUser = userService.findById(userId)
+            if (checkUser != null) {
+                return ResponseEntity(checkUser, HttpStatus.OK)
+            }
         }
         return ResponseEntity(ResponseMessage(false, "User not found"), HttpStatus.NOT_FOUND)
     }
 
-    @PutMapping("/user/{id}")
+    @PutMapping("/user")
     fun updateUserById(
-        @PathVariable(value = "id") userId: Long,
+        @RequestHeader("Authorization") token: String,
         @Valid @RequestBody updatedUser: User
     ): ResponseEntity<Any> {
-        val checkUser = userService.findById(userId)
-        if (checkUser != null) {
-            val user = checkUser.copy(
-                first_name = updatedUser.first_name,
-                last_name = updatedUser.last_name,
-                email = updatedUser.email,
-                password = hashService.hashBcrypt(updatedUser.password),
-                wallet = updatedUser.wallet
-            )
-            userService.save(user)
-            return ResponseEntity(ResponseMessage(true, "User updated successfully"), HttpStatus.OK)
+        val tokenValue = token.substring(7)
+        val userId = tokenService.getUserIdFromToken(tokenValue)
+        if (userId != null) {
+            val checkUser = userService.findById(userId)
+            if (checkUser != null) {
+                val user = checkUser.copy(
+                    first_name = updatedUser.first_name,
+                    last_name = updatedUser.last_name,
+                    email = updatedUser.email,
+                    password = hashService.hashBcrypt(updatedUser.password),
+                    wallet = updatedUser.wallet
+                )
+                userService.save(user)
+                return ResponseEntity(ResponseMessage(true, "User updated successfully"), HttpStatus.OK)
+            }
+            return ResponseEntity(ResponseMessage(false, "User not found"), HttpStatus.NOT_FOUND)
         }
         return ResponseEntity(ResponseMessage(false, "User not found"), HttpStatus.NOT_FOUND)
     }
 
-    @DeleteMapping("/user/{id}")
-    fun removeUserById(@PathVariable(value = "id") userId: Long): ResponseEntity<Any> {
-        val checkUser = userService.findById(userId)
-        if (checkUser != null) {
-            userService.deleteById(userId)
-            return ResponseEntity(ResponseMessage(true, "User deleted successfully"), HttpStatus.OK)
+    @DeleteMapping("/user")
+    fun deleteUserById(@RequestHeader("Authorization") token: String): ResponseEntity<Any> {
+        val tokenValue = token.substring(7)
+        val userId = tokenService.getUserIdFromToken(tokenValue)
+        if (userId != null) {
+            val checkUser = userService.findById(userId)
+            if (checkUser != null) {
+                userService.deleteById(userId)
+                return ResponseEntity(ResponseMessage(true, "User deleted successfully"), HttpStatus.OK)
+            }
+            return ResponseEntity(ResponseMessage(false, "User not found"), HttpStatus.NOT_FOUND)
         }
         return ResponseEntity(ResponseMessage(false, "User not found"), HttpStatus.NOT_FOUND)
     }
